refactor: 티켓 예매 기능 TicketingSerivce로 분리
