+++ backend/src/test/java/com/festago/application/integration/TicketServiceIntegrationTest.java
@@ -2,40 +2,30 @@ package com.festago.application.integration;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doReturn;
 
 import com.festago.application.TicketService;
 import com.festago.domain.Festival;
 import com.festago.domain.FestivalRepository;
-import com.festago.domain.Member;
-import com.festago.domain.MemberRepository;
-import com.festago.domain.MemberTicketRepository;
 import com.festago.domain.Stage;
 import com.festago.domain.StageRepository;
+import com.festago.domain.TicketAmount;
+import com.festago.domain.TicketAmountRepository;
+import com.festago.domain.TicketRepository;
 import com.festago.domain.TicketType;
 import com.festago.dto.TicketCreateRequest;
-import com.festago.dto.TicketingRequest;
-import com.festago.exception.BadRequestException;
+import com.festago.dto.TicketCreateResponse;
 import com.festago.exception.NotFoundException;
 import com.festago.support.FestivalFixture;
-import com.festago.support.MemberFixture;
 import com.festago.support.StageFixture;
 import java.time.Clock;
-import java.time.Instant;
 import java.time.LocalDateTime;
-import java.util.List;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.stream.IntStream;
+import java.time.ZoneOffset;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.boot.test.mock.mockito.SpyBean;
-import org.springframework.test.context.jdbc.Sql;
 
 @DisplayNameGeneration(ReplaceUnderscores.class)
 @SuppressWarnings("NonAsciiCharacters")
@@ -45,29 +35,29 @@ class TicketServiceIntegrationTest extends ApplicationIntegrationTest {
     TicketService ticketService;
 
     @Autowired
-    MemberRepository memberRepository;
-
-    @SpyBean
-    MemberTicketRepository memberTicketRepository;
-
-    @Autowired
     StageRepository stageRepository;
 
     @Autowired
     FestivalRepository festivalRepository;
 
-    @MockBean
+    @Autowired
+    TicketRepository ticketRepository;
+
+    @Autowired
+    TicketAmountRepository ticketAmountRepository;
+
+    @SpyBean
     Clock clock;
 
     @Test
     void 공연이_없으면_예외() {
         // given
-        String entryTime = "2023-07-26T18:00:00";
+        LocalDateTime entryTime = LocalDateTime.parse("2022-07-26T18:00:00");
         long invalidStageId = 0L;
         int totalAmount = 100;
 
         TicketCreateRequest request = new TicketCreateRequest(invalidStageId, TicketType.VISITOR,
-            totalAmount, LocalDateTime.parse(entryTime));
+            totalAmount, entryTime);
 
         // when && then
         assertThatThrownBy(() -> ticketService.create(request))
@@ -76,65 +66,59 @@ class TicketServiceIntegrationTest extends ApplicationIntegrationTest {
     }
 
     @Test
-    @Sql("/ticketing-test-data.sql")
-    void 동시에_100명이_예약() {
+    void 공연에_티켓_추가() {
         // given
-        int tryCount = 100;
-        Member member = memberRepository.save(MemberFixture.member().build());
-        TicketingRequest request = new TicketingRequest(1L);
-        ExecutorService executor = Executors.newFixedThreadPool(16);
-        doReturn(false)
-            .when(memberTicketRepository)
-            .existsByOwnerAndStage(any(Member.class), any(Stage.class));
-        doReturn(Instant.parse("2023-07-24T03:21:31Z"))
+        LocalDateTime stageStartTime = LocalDateTime.parse("2022-07-26T18:00:00");
+        doReturn(stageStartTime.minusWeeks(1).toInstant(ZoneOffset.UTC))
             .when(clock)
             .instant();
+        Festival festival = festivalRepository.save(FestivalFixture.festival()
+            .startDate(stageStartTime.toLocalDate())
+            .endDate(stageStartTime.toLocalDate())
+            .build());
+        Stage stage = stageRepository.save(StageFixture.stage()
+            .festival(festival)
+            .startTime(stageStartTime)
+            .ticketOpenTime(stageStartTime.minusDays(1))
+            .build());
+        TicketCreateRequest request = new TicketCreateRequest(stage.getId(), TicketType.VISITOR,
+            100, stageStartTime.minusHours(1));
 
         // when
-        List<CompletableFuture<Void>> futures = IntStream.range(0, tryCount)
-            .mapToObj(i -> CompletableFuture.runAsync(() -> {
-                ticketService.ticketing(member.getId(), request);
-            }, executor).exceptionally(e -> null))
-            .toList();
-        futures.forEach(CompletableFuture::join);
+        TicketCreateResponse response = ticketService.create(request);
 
         // then
-        assertThat(memberTicketRepository.count()).isEqualTo(50);
+        TicketAmount ticketAmount = ticketAmountRepository.findById(response.id()).get();
+        assertThat(ticketAmount.getTotalAmount()).isEqualTo(100);
     }
 
     @Test
-    @Sql("/ticketing-test-data.sql")
-    void 중복으로_티켓을_예매하면_예외() {
+    void 티켓이_있는_공연에_티켓을_추가하면_기존_티켓의_수량이_증가() {
         // given
-        Member member = memberRepository.save(MemberFixture.member().build());
-        TicketingRequest request = new TicketingRequest(1L);
-        Long memberId = member.getId();
-        doReturn(Instant.parse("2023-07-24T03:21:31Z"))
+        LocalDateTime stageStartTime = LocalDateTime.parse("2022-07-26T18:00:00");
+        doReturn(stageStartTime.minusWeeks(1).toInstant(ZoneOffset.UTC))
             .when(clock)
             .instant();
+        Festival festival = festivalRepository.save(FestivalFixture.festival()
+            .startDate(stageStartTime.toLocalDate())
+            .endDate(stageStartTime.toLocalDate())
+            .build());
+        Stage stage = stageRepository.save(StageFixture.stage()
+            .festival(festival)
+            .startTime(stageStartTime)
+            .ticketOpenTime(stageStartTime.minusDays(1))
+            .build());
+        TicketCreateRequest request = new TicketCreateRequest(stage.getId(), TicketType.VISITOR,
+            100, stageStartTime.minusHours(1));
 
-        ticketService.ticketing(memberId, request);
+        ticketService.create(request);
 
-        // when & then
-        assertThatThrownBy(() -> ticketService.ticketing(memberId, request))
-            .isInstanceOf(BadRequestException.class)
-            .hasMessage("예매 가능한 수량을 초과했습니다.");
-    }
+        // when
+        TicketCreateResponse response = ticketService.create(request);
 
-    @Test
-    void 티켓_생성시_입장_시간이_공연_시간보다_빠르면_예외() {
-        // given
-        Festival festival = festivalRepository.save(FestivalFixture.festival().build());
-        LocalDateTime now = LocalDateTime.now();
-        Stage stage = stageRepository.save(StageFixture.stage().festival(festival).startTime(now.plusHours(10))
-            .ticketOpenTime(now.plusHours(1)).build());
-        LocalDateTime entryTime = stage.getStartTime().plusHours(1);
-        Long stageId = stage.getId();
-        TicketCreateRequest request = new TicketCreateRequest(stageId, TicketType.STUDENT, 100, entryTime);
-
-        // when & then
-        assertThatThrownBy(() -> ticketService.create(request))
-            .isInstanceOf(BadRequestException.class)
-            .hasMessage("입장 시간은 공연 시간보다 빨라야합니다.");
+        // then
+        assertThat(ticketRepository.count()).isEqualTo(1);
+        TicketAmount ticketAmount = ticketAmountRepository.findById(response.id()).get();
+        assertThat(ticketAmount.getTotalAmount()).isEqualTo(200);
     }
 }
